Move Task classes to utils package
